refactor: TgLockDB -> TgDB
