+++ acceptance-tests/src/test/scala/io/renku/graph/acceptancetests/db/TriplesStore.scala
@@ -24,8 +24,8 @@ import eu.timepit.refined.auto._
 import io.renku.db.DBConfigProvider
 import io.renku.graph.model.{RenkuUrl, projects}
 import io.renku.projectauth.{ProjectAuthData, QueryFilter}
-import io.renku.triplesgenerator.TgLockDB.SessionResource
-import io.renku.triplesgenerator.{TgLockDB, TgLockDbConfigProvider}
+import io.renku.triplesgenerator.TgDB.SessionResource
+import io.renku.triplesgenerator.{TgDB, TgLockDbConfigProvider}
 import io.renku.triplesstore._
 import io.renku.triplesstore.client.util.JenaRunMode
 import org.typelevel.log4cats.Logger
@@ -37,7 +37,7 @@ object TriplesStore extends InMemoryJena with ProjectsDataset with MigrationsDat
 
   protected override val jenaRunMode: JenaRunMode = JenaRunMode.FixedPortContainer(3030)
 
-  private val dbConfig: DBConfigProvider.DBConfig[TgLockDB] =
+  private val dbConfig: DBConfigProvider.DBConfig[TgDB] =
     new TgLockDbConfigProvider[Try].get().fold(throw _, identity)
 
   lazy val sessionResource: Resource[IO, SessionResource[IO]] =
