+++ triples-generator/src/main/scala/io/renku/triplesgenerator/events/consumers/cleanup/SubscriptionFactory.scala
@@ -26,7 +26,7 @@ import io.renku.events.consumers.subscriptions.SubscriptionPayloadComposer.defau
 import io.renku.graph.model.RenkuUrl
 import io.renku.metrics.MetricsRegistry
 import io.renku.triplesgenerator.Microservice
-import io.renku.triplesgenerator.TgLockDB.TsWriteLock
+import io.renku.triplesgenerator.TgDB.TsWriteLock
 import io.renku.triplesgenerator.events.consumers.tsmigrationrequest.migrations.reprovisioning.ReProvisioningStatus
 import io.renku.triplesstore.{ProjectSparqlClient, SparqlQueryTimeRecorder}
 import org.typelevel.log4cats.Logger
