+++ triples-generator/src/test/scala/io/renku/triplesgenerator/events/consumers/syncrepometadata/EventHandlerSpec.scala
@@ -30,7 +30,7 @@ import io.renku.generators.Generators.Implicits._
 import io.renku.graph.model.projects
 import io.renku.interpreters.TestLogger
 import io.renku.lock.Lock
-import io.renku.triplesgenerator.TgLockDB.TsWriteLock
+import io.renku.triplesgenerator.TgDB.TsWriteLock
 import io.renku.triplesgenerator.api.events.Generators.syncRepoMetadataEvents
 import io.renku.triplesgenerator.events.consumers.TSReadinessForEventsChecker
 import io.renku.triplesgenerator.events.consumers.syncrepometadata.processor.EventProcessor
